chore: Remove duplicate definitions
